recurse on directories
If a directory name is given as an argument, we currently get an error
saying "file not found".

Also the only way to recurse on a set of passwords is currently to
report on all passwords.

It would be nice to be able to report only on a subset of the repository
tree. For this, if a directory name is passed in as an argument, we
should recurse over all ".gpg"-ending files contained within this
directory.